Reduce overhead of sanity checks on compiler startup
Avoid putting he llvm version in the sanity file, since the cost
of running `clang --version` is fairly significant.  On my machine this
was about 25% of the `emcc` overhead (compared to just calling clang).

We will check the llvm version but only at link time, with the idea
the number of link commands is dwarfed by the number of compile
commands.

On my machine this saved about 5% on `./embuilder build libc --force`

See: #18623