Add a port of mimalloc, a fast and scalable multithreaded allocator  (#20651)
The new allocator can be used with -sMALLOC=mimalloc.

On the benchmark added in this PR, dlmalloc does quite poorly here (getting
actually slower with each additional core, because the lock contention is much
larger than the actual work in the artificial benchmark). mimalloc, in
comparison, scales the same as natively: more cores keeps helping. So mimalloc
can be a significant speedup in codebases that have lock contention on malloc.

mimalloc is significantly larger than dlmalloc, however, so we do not want it
on by default. It also uses more memory, because of how mimalloc works and also
due to #20645.

Design-wise, this layers mimalloc on top of emmalloc. emmalloc functions as the
"system allocator", which is more powerful than just using raw sbrk - sbrk can't
free holes in the middle, for example.

Code-wise, all of system/lib/mimalloc is unchanged from upstream (see
README.emscripten) except for an ifdef or two, and then the new backend which
is in system/lib/mimalloc/src/prim/emscripten/prim.c. That file has more
comments explaining the design of the port.

A new test is added which is also usable as a benchmark,
test/other/test_mimalloc.cpp, which is where the numbers above come from.

Fixes #18369