split linker code into a separate file. NFC (#20717)
This change moves more than half of the code in emcc.py which is used
only for linking into its own file (link.py).  This makes reasoning
about the code easier and paves the way for a fast path then only
compiling (e.g. this code shouldn't even need to be imported when only
compiling).  See #20577.