Reduce overhead of sanity checks on compiler startup (#20734)
Stop included the llvm version in the sanity file.  The  the cost
of running `clang --version` is fairly significant.  On my machine this
was about 25% of the `emcc` overhead (compared to just calling clang).

With this change we no longer check the llvm version when compiling
and only when linking.  The idea here is that for most project link
commands are more rare that compile commands.

The slight downside is that we won't catch folks using the wrong version
of clang until they get to link time.

See: #18623