Remove warning about very large functions.
4 years ago, commit 49d7717600bdb89010e84f50bf326a34ee85d439 removed
OUTLINING_LIMIT which was the only actionable item a user could do to avoid
this warning. Suggesting to compile with -Oz when the user is already
compiling with -Oz is just stupidly insulting and not constructive user
feedback.

It is not Emscripten's job to have opinions about the size of functions.

Fixes <https://github.com/emscripten-core/emscripten/issues/20811>.