Streamline metadce by avoiding wasted work (#20814)
Clarify that we only remove imports and exports, and use that information to
only call the "apply" stage if we in fact have something to remove. In practice
it looks like we end up with nothing to remove only when using the minimal
runtime, as the normal runtime has stuff like errno that is almost always
removable.

This is probably not a significant speedup, but it doesn't hurt, and it is nice I
think to clarify what work we do exactly.