Remove some magic handling of native shared libraries
We had this rather strange (I imagine completely unused) feature where
is emscripten is passed a concrete path to a shared library file (e.g.
/path/to/libfoo.so) and it could determine that the shared library was
not indented for emscripten then it would instead search the library
for other things called `libfoo` that might work.

This seems too magical at best and harmful/confusing a worst.

The presence of this feature is blocking the way for some internal
optimizations and refactorings (e.g. #20884)