Use babel for transpiling rather than closure compiler (#20879)
We recently switch from using closure-compiler with `WHITESPACE_ONLY` to
closure-compiler with `SIMPLE_OPTIMIZATIONS`.  However this had the
negative side effect that all input need to be free of closure compiler
warnings, and this turned out not to be practical for all users.
See #20810 for more on this

When selecting a transpilation tool use I also evaluated `swx` (written
in rust) and `esbuild` (written in go).  `esbuild` was rejected
because the simply don't support ES5
(https://github.com/evanw/esbuild/issues/297).  `swx` was rejected
because it almost doubled the side of our `node_modules` directory by
adding two 50mb binary files.