Remove if/loop tests, and a fastcomp test (#20897)
Removes the silliest tests we had: a single loop, a single if, and an if-else.

Also remove test_multiply_defined_symbols which only made sense in
fastcomp. There are no actually multiply defined symbols in that test, aside
from during the link stage, so this tested that the fastcomp .a linker handled
that properly. But lld does that today, and does not need testing here. Even
if we did want to test it, this is a poor test as the duplicate symbols are not
referred to anyhow, so they are not even collected from the respective .a
files.