Remove support for --default-obj-ext
As far as I can tell the only user of this option was add mostly in
support for custom cmake and visual studio integration that was removed
in #20901.

This feature is something that neither clang nor gcc and I'd like to
remove it so that we can move closer to being a pure clang wrapper.