Add switch to disable batching of inputs when building system libs
Basically gives emscripten developers a way to opt out of #20577.

I've found it useful for debugging to build just one file at time and
have absolute (not relative) paths in the build commands.  This allows
me to copy and paste a single failing command and run it in isolation
when working on system library changes.