Create a setting reference in the docs
This should be able to replace https://emsettings.surma.technology/

A big upside here is that we a link that we can use for each setting
as well as the ability to link to a given setting internally.

As part of this change I also updated the settings docs themselves to be
formatting in a ReSt-friendly way.

I also had to rename the section tag for `wasm_workers.rst` and
`asyncify.rst` because the happen to collide with the ref for the
corresponding setting name.

Fixes: #16563