Remove yet more trivial testing (#20964)
Remove trivial test for an "empty" class (with no properties).

Remove a test for a name collision between a function parameter name and a
libc function. That was a big deal back in 2011 when emscripten parsed LLVM IR
to find names for things, but now even in wasm2js we do not use param names
from C++ source.

Dynamic cast tests are C++ library tests, and so other is enough for them.

Run function pointer tests only in wasm2js. This is basic clang stuff, but still, it is
good to have some coverage for wasm2js on them I think - that it emits the table
right and does not misoptimize them etc.