Remove trivial testing of stack usage (#21001)
All these tests do is allocate (or try to, at least) many things on the stack, either
using structs passed by value or printf. But this is simply a clang codegen issue
now, so we can leave it to there. Also, some of these tests were not even valid:
they almost tested whether we used an unexpectedly large amount of stack,
but they didn't do even that as the removed comments about a larger stack
size show (i.e. we enlarged the stack size to work around a CI issue, but that
meant we aren't testing that we work even with a small stack size).

Regardless we have good testing for our stack checks, which are the proper
solution to this issue (that is, we don't break weirdly if the user set the stack
as too small), and so we don't need to duplicate any clang testing.