Use musl's strptime implementation
This seems to save on code size compared the JS version as well as
making our Wasm files more standalone.  I measured the combined size of
core2.test_strptime_days (with closure compiler enabled) and got the
following results:

- before: 17113 bytes
- after: 16398 bytes

As part of this I had to fix several missing part of the musl's
implementation.  We also loose support for "%U" and "%W"
specifiers for week, which is not supported by either bionic or musl
upstream. See:

- https://github.com/aosp-mirror/platform_bionic/blob/4cfb2bde90fcf0a1f835e2b670b00e9f226c1c7c/libc/tzcode/strptime.c#L364-L375
- https://github.com/bminor/musl/blob/f314e133929b6379eccc632bef32eaebb66a7335/src/time/strptime.c#L126-L132

In both cases these values get ignored, so it seems reasonable that
emscripten do that same.