Don't add `Module.inspect` when building for node
This was added back in 2015 but without any rationale.  See 8d4d62eefc.

My guess is that this was somehow related to node's `util.inspect`
function but this doesn't seem to be the way to work with that.
We would need to do something more like this to make that work:
https://nodejs.org/api/util.html#custom-inspection-functions-on-objects

Perhaps it has some other purpose but I have no idea what that might be.