[EH] Check if Exception.stack is writable
Apparently `WebAssembly.Exception.stack` has been `readonly` all along
and I wasn't aware. But I wonder why this has to be `readonly`, and
currently writing to this property works in Chrome and Node and fails in
Firefox. I'll ask in the EH repo whether this should be `readonly`, but
in the meantime, we check if this is writable before writing to it.

Temporarily fixes #21044.

Not sure how to test this... Do we run Firefox tests?