Fix dropping of arguments in JS library functions
When the `__sig` for a function is shorter than the actual number of
argument (i.e. when the native signature is shorter than the JS
signature) we were dropping the extra arguments in handleI64Signatures.

Specifically the `uuid_unparse` library function has a hidden third
argument that is not exposed to native code (doesn't appear in the
native signature).

Fixes: #21056