[EH] Don't write to readonly Exception.stack property
Apparently `WebAssembly.Exception.stack` has been `readonly` all along
and I wasn't aware. While I'm not sure if this should be `readonly` and
this may change in the future, for now writing to this property is not
spec compliant, so this removes the routine that tidies up the stack
trace, which is not the end of the world.

Fixes #21044.