Re-enable running of closure compiler under java (#21093)
This is a partial revert of #20919.

It turns out that we were still depending on falling back to the java
version of closure compiler in some cases.  Specifically, on macOS arm64
when Rosetta is not installed, we were falling back to the Java version
since google-closure-compiler doesn't ship arm64 binaries.  I imagine
the same will be true of linux-arm64.