Remove remaining uses of setErrNo and deprecate SUPPORT_ERRNO setting (#21074)
Aside from the `setErrNo` library function `SUPPORT_ERRNO` also
controlled the building the malloc and sbrk.  However, just setting
errno to ENOMEM didn't seem to have any effect on the code size for any
of our micro benchmarks.  This may be because LTO can completely
eliminate the write to `errno` when there are no readers, or it could be
that errno usage exists in in even the smallest program already so
removing from malloc alone makes no difference.