[ci] Update to latest version of cicleci windows orb (#21110)
We got a message saying that the version vs2019 will be deprecated this
year.  This change doesn't stop using vs2019, but just updates to the
latest version of the windows orb.  We may still need to move away from
vs2019 in the future, but if possible we would like to continue testing
on older version of windows so that we can be sure we continue to
support them.