Avoid use of `lsan::GetCurrentThread` in `lsan_common_emscripten.cpp`
This code was depending in a dummy version of `lsan::GetCurrentThread`
that was added to `asan_emscripten.cpp`.

However as far as I can tell this never worked since the dummy version
of `lsan::GetCurrentThread` was returning a TID directly, but its
expected to return a pointer.

I ran some logging and saw the following result when running under asan
(i.e. when the dummy function was being used):

`GetCurrentThreadId=ffffffff GetCurrentThread=0x00000000.`