file_packager.py: Add --optional
If `--optional` is specified, the generated JS will remove all run dependencies when encountering an error for which no fallback is available (ex. a fetch `.data` error), so that running is unblocked.

If not specified, the generated JS will continue to block running, and will throw an Error (the same as previous behavior).