Add reference type logging support
We need these if we want to enable reference-types feature in Clang by
default.

`test_autodebug_wasm` fails without this if we enable reference-types in
Clang. (Even though the test itself doesn't seem to actually use them,
turning on `EMCC_AUTODEBUG` runs `wasm-opt` with `--instrument-locals`,
which adds imports for `get_` and `set_` functions for the reference
types if reference-types feature is enabled:
https://github.com/WebAssembly/binaryen/blob/8c834e8257b03ea87b639ddac9adefec64fcad00/src/passes/InstrumentLocals.cpp#L179-L187