Generate TypeScript definitions for runtime exports.
When building the JS library, collect the JS and corresponding JS docs so
they can be used by the TypeScript compiler to generate definitions for
any of the exported runtime functions.

I also added a test for using the output in a TS file.

Note: Initially I tried running tsc on the final JS output, but it
often got confused when the Module had dynamic properties added to it.
Maybe once we clean things up this will be a better option.