Finish converting JS compiler to ES6 modules. NFC (#21542)
The compiler itself is now purely JS modules.

As part of this change I create a new global context object in which
all library code is processed.  This allows is to have a clean
separation between the compiler code itself and code being processed.
Only symbols explicitly added to the macro context are available to JS
library code.