Avoid exporting "ready" promise on the module object (#21564)
I believe the "ready" promise was set on the Module object here
primarily so that closure would not minify the name.  Instead of
polluting the Module namespace we can instead just tell closure not to
minify it.

In addition to polluting the `Module` namespace exporting a `ready`
promise is also confusing since the only way to get a module object when
`-sMODULARIZE` is used is via waiting on the promise returned from the
factory function.  So, by definition, the promise has already resolved
before anyone can access it.

Instead we use the `closure-externs.js` file to suppress the minifiction
which matches the behavior of the incoming `moduleArg` argument.