Fix for zero count in glUniform family of functions
The previous fix for this was in #16837, but I looks like that only
covered the new "garbage-free" webgl2 path.  When old webgl1 path was
still using the zero count value.

As part of this change I resurrected
`test_webgl_draw_triangle_with_uniform_color.c` which has not actually
be used since #20925.

Fixes: #21567