JS type legalized is still performed regardless of LEGALIZE_JS_FFI
`wasm-emscripten-finalize` will always performs some amount of type
legalization unless WASM_BIGINT is enabled.  The logic in
`finalize_wasm` was getting this wrong.

I'm not sure exactly what the side effects of this are today but I'm
working on change to `getTempRet0`/`setTempRet0` handling where this
becomes a problem because `js_legalization_pass_flags` was not being
passed to `wasm-emscripten-finalize` even though some amount of
legalization was happening.  This was only effecting the `WASM_BIGINT=0`
+ `LEGALIZE_JS_FFI=0` case in which partial legalization occurs.

See: #21579