[embind] Only allow class types for unique_ptr bindings.
As pointed out by #21625, when using unique_ptr's with primitive types the
raw pointer was being passed between wasm and JS, but it was using the
underlying type (e.g. char, int, etc) to convert the value. This means the
pointer was truncated to a char at the boundary.

Embind doesn't allow pointers or references to primitive types so it doesn't
make sense to allow a unique_ptr for these types either.