Fix destructor return type in __cxa_init_primary_exception
https://github.com/llvm/llvm-project/pull/65534 introduces
`__cxa_init_primary_exception` and uses this in libcxx. Like several
other methods like the below in `cxa_exception.cpp`,
https://github.com/emscripten-core/emscripten/blob/fbdd9249e939660cb0d20f00d6bc1897c2f3905e/system/lib/libcxxabi/src/cxa_exception.cpp#L264-L269
this new function takes a pointer to a destructor, and in Wasm
destructor returns a `this` pointer, which requires
`ifdef __USING_WASM_EXCEPTION__` on its signature.

And that it is also used in libcxx means we need to guard some code in
libcxx with `ifdef __USING_WASM_EXCEPTION__` for the signature
difference, and we need to build libcxx with
`-D__USING_WASM_EXCEPTIONS__` when Wasm EH is used.