Disable hardening mode
LLVM 18.1.2 adds support for "Hardening Modes" to libcxx
(https://libcxx.llvm.org/Hardening.html). There are four modes: none,
fast, extensive and debug. Different hardening modes make different
trade-offs between the amount of checking and runtime performance. We
for now disable it (i.e., set it to 'none') so that we don't have any
effects on the performance. We can consider enabling it when we ever get
to enable the debug version of libcxx.