Disable __libcpp_verbose_abort directly in code
In #20707 we decided to disable `__libcpp_verbose_abort` not to incur
the code size increase that brings (it brings in `vfprintf` and its
family). We disabled it in adding
`#define _LIBCPP_AVAILABILITY_HAS_NO_VERBOSE_ABORT` in `__config_site`.

That `_NO_` macros are gone in LLVM 18.1.2, so I changed it to
`#define _LIBCPP_AVAILABILITY_HAS_VERBOSE_ABORT 0`
https://github.com/emscripten-core/emscripten/pull/21638/commits/2bd7c670b6ab9ef7c89cf68122bfc239596d1746
But apparently, when including `__availability`, that setting is
overridden by this line:
https://github.com/llvm/llvm-project/blob/38f5596feda3276a8aa64fc14e074334017088ca/libcxx/include/__availability#L138
because we don't define our own vendor availability annotations
https://github.com/emscripten-core/emscripten/blob/193b2bff980529dbf5d067f01761ed8a6e41189f/system/lib/libcxx/include/__config_site#L4

I asked about this in
https://github.com/llvm/llvm-project/issues/87012 and
https://github.com/llvm/llvm-project/pull/71002#issuecomment-2024604569
(and more comments below)
but didn't get an actionable answer yet.

This disables `__libcpp_verbose_abort` in the code directly for now,
hopefully temporarily.