Ensure ENVIRONMENT_IS_PTHREAD/WASM_WORKER are defined before pre-js file
Its useful for pre-js files to be able to depend on these variables,
and this is specifically necessary for #21701 which removes the default
export of `ENVIRONMENT_IS_PTHREAD`.

Note that this also fixes an inconsistency between the regular runtime
and the minimal runtime where minimal runtime was defining some
`ENVIRONMENT_IS_XX` var before pre-js but the regular runtime was not.