Handle the case where wasm data dependency contains special character `@`
In this special case, `src@dst` could look like `xxxx@v1@xxxx@v1`

If we break by first `@` position, then the path will be `xxxx` which result in file not found error (since its supposed to be `xxxx@v1`)

If the data dependency contains `@`, then there will be odd number of occurence for sure. the middle occurence must be the correct split.