Improve testing of `mainScriptUrlOrBlob`. NFC
I noticed when working on #21701 then these tests were not renaming
the `.js` output, which meant that if `mainScriptUrlOrBlob` was
completely ignored the tests would still pass.

These changes rename the output `.js` file such that the default
name will not longer work and the tests would fail if
`mainScriptUrlOrBlob` were ignored.