Don't run `--extern-pre-js` or `--extern-post-js` code on pthreads
This means, for example, that folks can call the module constructor as
part of `--extern-post-js` without accidentally creating an extra module
instance on each thread.

For an example of this pattern see
test_pthread_offset_converter_modularize.

Although this is minor code size regression I'm about to followup with
a major refactor that saves a lot more code: #21701