Remove internal emscripten_get_module_name helper. NFC
This helper was added when we implement sanitizes in #8651.  It was
never added to any public header file and should only have these two
internal users.

Instead of this function we can just use the existing internal
`_emscripten_get_progname` function, which can in turn then be
implemented in terms of the existing `getExecutableName` JavaScript
function.