Skip wasmBinaryFile initialization on pthread workers
Delay the search for the wasm binary (e.g. using `locateFile`) so that
it does not run on pthread workers (or weasm workers) where
`Module.instantiateWasm` is set.

The wasm file is always received via postMessage when running on the
worker so we shouldn't ever be calling locateFile in this situation.

See #21826 and #21827