Fix pthread + modularize + running in node worker
The issue here was the special case for the setting of `scriptDirectory`
under nodejs when `ENVIRONMENT_IS_WORKER`.

Previously this path was not being hit when the main thread was running
in a worker because we were not setting `ENVIRONMENT_IS_WORKER` in this
case.  This was fixed in #21701 and `ENVIRONMENT_IS_WORKER` was
correctly set, but that meant that `scriptDirectory` then started being
set incorrectly.

Fixes: #21827