Remove regal port
The regal project seems like its been dead for many years and the repo
from which it was forked doesn't seem to exist anymore:
https://github.com/p3/regal

I'm not sure we want to continue supporting it as first class thing if
there is nobody upstream to report issues to.  If it turns out that
there are folks who really depend on this we can consider moving it into
`tools/ports/conrib`.

See #21885