[SjLj] __USING_WASM_SJLJ__ -> __WASM_SJLJ__
In line with #21970, this changes `__USING_WASM_SJLJ__` to
`__WASM_SJLJ__`. We don't define this in Clang now (which we can maybe
consider, but given that Clang doesn't have dedicated flags for SjLj and
there don't seem to be other platforms doing similar things), but given
that the affected files currently are only in Emscripten, I think we can
safely change this.