[EH] Guard destructor changes in libc++abi with __wasm__
These changes are due to that Wasm destructor returns its argument, so
in case someone wants to compile out libc++abi _without_ using Wasm EH,
these will error out too. In Emscripten this has been fine because in
`system_lib.py` we only build these files when Wasm EH is enabled, but
this may not be the case if we upstream our changes.