[asyncify] Refactor argument restoring under wasm64. NFC
This change was split out from my larger change to switch to table64.

When we use MEMORY64 + ASYNCIFY=1 we have add multiple wrappers to the
wasm exports.  This change basically moves the argument restoration to
the outside rather than doing int the wrapper.  This means that all the
different wrappers see the restored arguments, not just the asyncify
wrapper (which is the first/lowest level wrapper.