[tsgen] Check parameter types match
When using embind::function() with --emit-tsd, parameter names for TypeScript definitions cannot be detected automatically and must be supplied manually in string form in the function name. This is error-prone as it duplicates the function prototype but in a form that the compiler can't type-check. This change adds some macros which allow specifying the parameter names and types in a form that the compiler can type-check. This will reduce the likelihood of parameters in embind::function() declarations getting out of sync with the functions they expose.

This change also reduces the burden of maintaining the parameter names, as one can simply copy and paste directly from the function prototype without modification. Previously you were required to edit out the types from the function prototype manually.